+++ cookbooks/couchbase/metadata.rb
@@ -4,7 +4,7 @@ maintainer_email "jdunn@opscode.com"
 license          "MIT"
 description      "Installs and configures Couchbase Server."
 long_description IO.read(File.join(File.dirname(__FILE__), "README.md"))
-version          "2.0.0"
+version          "2.0.1"
 
 %w{debian ubuntu centos redhat oracle amazon scientific windows}.each do |os|
   supports os