+++ cookbooks/java/recipes/oracle-7.rb
@@ -1,4 +1,5 @@
 yum_package "jdk.x86_64" do
   flush_cache [ :before ]
-  action [ :install, :upgrade ]
+  action [ :install ]
 end
+