+++ cookbooks/webplayer/recipes/default.rb
@@ -58,9 +58,9 @@ end
 memcached_servers = Array.new
 res.each do |s|
   if node.chef_environment == "prod"
-    memcached_servers << s[:hostname] + "-v200.ihr:11212"
+    memcached_servers << s[:hostname] + "-v200.ihr:11211"
   else
-    memcached_servers << s[:hostname] + ".ihr:11212"
+    memcached_servers << s[:hostname] + ".ihr:11211"
   end
 
 end